face_chat_api/requirements.txt
fastapi
uvicorn
facenet-pytorch
torchvision
torch
scipy
numpy
pillow
requests
python-multipart
lancedb